+++ docs/.vitepress/sidebar.ts
@@ -15,6 +15,7 @@ export const mainSidebar: DefaultTheme.SidebarItem[] = [
     text: "Features",
     items: [
       { text: "Gates", link: "/features/gates" },
+      { text: "Epilogues", link: "/features/epilogues" },
       { text: "Arguments", link: "/features/args" },
       { text: "Command tree", link: "/features/command-tree" },
       { text: "Plugins", link: "/features/plugins" },